refact:#43 component - layout 패키지 분리, 공통 component 분리
