+++ app/src/main/java/com/dev/briefing/presentation/chat/ChattingScreen.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.dev.briefing.BuildConfig
-import com.dev.briefing.presentation.setting.CommonHeader
+import com.dev.briefing.util.component.CommonHeader
 import com.google.accompanist.web.AccompanistWebChromeClient
 import com.google.accompanist.web.AccompanistWebViewClient
 import com.google.accompanist.web.WebView
