+++ app/src/main/java/com/dev/briefing/presentation/setting/SettingScreen.kt
@@ -1,51 +1,37 @@
 package com.dev.briefing.presentation.setting
 
 import android.app.*
-import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import android.util.Log
 import android.widget.Toast
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.core.app.ComponentActivity
-import androidx.core.app.NotificationCompat
 import androidx.core.content.ContextCompat.startActivity
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.dev.briefing.BuildConfig.NOTIFICATION_CHANNEL_ID
 import com.dev.briefing.R
 import com.dev.briefing.presentation.login.SignInActivity
 import com.dev.briefing.presentation.login.SignInViewModel
+import com.dev.briefing.presentation.setting.component.menuWithArrow
 import com.dev.briefing.presentation.theme.*
 import com.dev.briefing.presentation.theme.utils.CommonDialog
-import com.dev.briefing.util.ALARM_CODE
 import com.dev.briefing.util.ALARM_TAG
 import com.dev.briefing.util.JWT_TOKEN
 import com.dev.briefing.util.MEMBER_ID
 import com.dev.briefing.util.MainApplication.Companion.prefs
 import com.dev.briefing.util.REFRESH_TOKEN
-import com.dev.briefing.util.SharedPreferenceHelper
+import com.dev.briefing.util.component.CommonHeader
+import com.dev.briefing.util.extension.convert.formatTime
 import org.koin.androidx.compose.getViewModel
 import org.koin.androidx.compose.koinViewModel
-import java.text.SimpleDateFormat
-import java.util.*
 
 @Composable
 fun SettingScreen(
@@ -98,7 +84,6 @@ fun SettingScreen(
             dialogText = R.string.dialog_exit_text,
             dialogId = R.string.dialog_exit_confirm
         )
-
     }
     if (openLogOutDialog.value) {
         CommonDialog(
@@ -118,8 +103,6 @@ fun SettingScreen(
             dialogText = R.string.dialog_logout_text,
             dialogId = R.string.dialog_logout_confirm
         )
-
-
     }
     LazyColumn(
         modifier = modifier
@@ -229,138 +212,6 @@ fun SettingScreen(
                 openExitDialog.value = true
             }, color = DialogExit)
             Spacer(modifier = Modifier.height(100.dp))
-
         }
-
-
-    }
-}
-
-fun formatTime(
-    hour: Int,
-    minute: Int,
-): String {
-    var tmpString = ""
-    if (hour <= 12) {
-        tmpString = "오전 " + "${hour}시 ${minute}분"
-    } else {
-        tmpString = "오후 " + "${hour - 12}시 ${minute}분"
-    }
-    return tmpString
-}
-
-@Composable
-fun menuWithText(
-    @StringRes menu: Int = R.string.navigation_chat,
-    onClick: () -> Unit = {},
-    color: Color = MainPrimary2,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(color = White, shape = RoundedCornerShape(5.dp))
-            .padding(horizontal = 12.dp, vertical = 12.dp)
-            .clickable(onClick = onClick),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = stringResource(id = menu),
-            style = MaterialTheme.typography.titleSmall.copy(
-                fontWeight = FontWeight(400),
-                color = color
-            )
-        )
-    }
-}
-
-@Composable
-fun menuWithArrow(
-    time: String = "",
-    isArrow: Boolean = true,
-    @DrawableRes icon: Int = R.drawable.setting_caution,
-    @StringRes menu: Int = R.string.navigation_chat,
-    onClick: () -> Unit = {},
-    modifier: Modifier = Modifier
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(color = White, shape = RoundedCornerShape(5.dp))
-            .padding(horizontal = 12.dp, vertical = 12.dp)
-            .clickable(onClick = onClick),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Row {
-            Image(
-                painter = painterResource(id = icon),
-                contentDescription = "더보기"
-            )
-            Spacer(modifier = Modifier.width(12.dp))
-            Text(
-                text = stringResource(id = menu),
-                style = MaterialTheme.typography.titleSmall.copy(
-                    fontWeight = FontWeight(400)
-                )
-            )
-        }
-        if (isArrow) {
-            Image(
-                painter = painterResource(id = R.drawable.left_arrow),
-                contentDescription = "더보기"
-            )
-        } else {
-            Text(
-                text = time,
-                style = MaterialTheme.typography.titleSmall.copy(
-                    fontWeight = FontWeight(400),
-                    color = MainPrimary3
-                )
-            )
-        }
-
-
-    }
-}
-
-@Composable
-fun CommonHeader(
-    onBackClick: () -> Unit,
-    header: String = "",
-    color: Color = SubBackGround
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(color = color)
-            .padding(top = 60.dp, bottom = 20.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    )
-    {
-        Image(
-            painter = painterResource(
-                id = R.drawable.vector
-            ),
-            contentDescription = "뒤로가기", modifier = Modifier
-                .clickable(onClick = onBackClick)
-        )
-        Text(
-            text = header,
-            style = MaterialTheme.typography.titleMedium.copy(
-                color = MainPrimary,
-                fontSize = 24.sp,
-                fontWeight = FontWeight(400)
-            )
-        )
-        Text(
-            text = "",
-            style = MaterialTheme.typography.titleMedium.copy(
-                color = MainPrimary
-            )
-        )
-
     }
 }
