+++ app/src/main/java/com/dev/briefing/util/component/CommonHeader.kt
@@ -1,0 +1,62 @@
+package com.dev.briefing.util.component
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.dev.briefing.R
+import com.dev.briefing.presentation.theme.MainPrimary
+import com.dev.briefing.presentation.theme.SubBackGround
+
+@Composable
+fun CommonHeader(
+    onBackClick: () -> Unit,
+    header: String = "",
+    color: Color = SubBackGround
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(color = color)
+            .padding(top = 60.dp, bottom = 20.dp),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    )
+    {
+        Image(
+            painter = painterResource(
+                id = R.drawable.vector
+            ),
+            contentDescription = "뒤로가기", modifier = Modifier
+                .clickable(onClick = onBackClick)
+        )
+        Text(
+            text = header,
+            style = MaterialTheme.typography.titleMedium.copy(
+                color = MainPrimary,
+                fontSize = 24.sp,
+                fontWeight = FontWeight(400)
+            )
+        )
+        Text(
+            text = "",
+            style = MaterialTheme.typography.titleMedium.copy(
+                color = MainPrimary
+            )
+        )
+
+    }
+}
